Cleaned up the rust models.
