--- Models/gate-layer/src/gate-internals/Cargo.toml
@@ -1,10 +1,0 @@
-
-[package]
-name = "gate_internals"
-version = "0.1.0"
-edition = "2021"
-
-
-[lib]
-name = "gate_internals"
-path = "lib.rs"
