+++ Models/utilities/src/gate_internals.rs
@@ -1,4 +1,3 @@
-
 pub fn GetBoolVector<const  BIT_WIDTH: usize>(value: usize) -> [bool; BIT_WIDTH] {
     let mut bit_store = [false; BIT_WIDTH];
     for i in 0..BIT_WIDTH{
@@ -82,48 +81,3 @@ impl Base2GateControlFunc {
         func ^ c
     }
 }
-
-macro_rules! create_wire_enum {
-    ($name:ident<$output:ty> { $($variant:ident = $value:expr),* $(,)? }) => {
-        #[derive(Debug, Clone, PartialEq, Eq)]
-        pub enum $name {
-            $($variant),*
-        }
-
-        impl $name {
-            pub fn all_variants() -> Vec<Self> {
-                vec![$(Self::$variant),*]
-            }
-
-            pub fn get_value(&self) -> $output {
-                match self {
-                    $(Self::$variant => $value as $output),*
-                }
-            }
-            
-            pub fn value_from(val: $output) -> Self {
-                match val {
-                    $($value => Self::$variant,)*
-                    _ => unreachable!()
-                }
-            }
-        }
-    };
-}
-
-create_wire_enum! {
-    Wires<u8>{
-        Wire1 = 1,
-        Wire2 = 2,
-        Wire3 = 3,
-        Wire4 = 4,
-        Wire5 = 5,
-        Wire6 = 6,
-        Wire7 = 7,
-        Wire8 = 8,
-        Wire9 = 9,
-        Wire10 = 10,
-        Wire11 = 11,
-    }
-}
-
