+++ Models/utilities/src/lib.rs
@@ -1,0 +1,16 @@
+use std::io::Result;
+use std::{fs::File, io::Write, usize};
+pub mod gate_internals;
+pub mod wires;
+
+pub fn write_file(file_path: &str, contents: &str) -> Result<()> {
+    let mut file = File::create(file_path)?;
+    file.write_all(contents.as_bytes())
+}
+
+
+pub trait Layer<const NUMBER_OF_WIRES: usize> {
+    type Items;
+
+    fn evaluate(&self, values: [Self::Items; NUMBER_OF_WIRES]) -> [Self::Items; NUMBER_OF_WIRES];
+}
